requirements.txt
absl-py==2.4.0
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==25.4.0
backoff==2.2.1
bitsandbytes==0.49.2
blis==1.3.3
boto3==1.42.89
botocore==1.42.89
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
confection==1.3.3
contourpy==1.3.3
coverage==7.13.4
cryptography==46.0.7
cycler==0.12.1
cymem==2.0.13
dill==0.4.1
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
fastapi==0.135.3
filelock==3.25.2
fonttools==4.62.0
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.1
google-pasta==0.2.0
grad-cam==1.5.5
graph_of_thoughts==0.0.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llmlingua==0.2.2
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.19
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.4
numba==0.65.0
numpy==2.4.2
omegaconf==2.3.0
openai==2.31.0
opencv-python==4.13.0.92
packaging==24.2
pandas==3.0.1
pathos==0.3.5
phonenumbers==9.0.28
pillow==12.1.1
platformdirs==4.9.6
pluggy==1.6.0
pox==0.3.7
ppft==1.7.8
preshed==3.0.13
presidio_analyzer==2.2.359
presidio_anonymizer==2.2.362
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
sagemaker==2.257.1
sagemaker-core==1.0.77
schema==0.7.8
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.6.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
sympy==1.14.0
tblib==3.2.2
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.22.2
torch==2.11.0
torch-tb-profiler==0.4.3
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.3
ttach==0.0.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.44.0
wasabi==1.1.3
weasel==1.0.0
websockets==16.0
Werkzeug==3.1.8
wrapt==2.1.2
zipp==3.23.1
psycopg2==2.9.11
pywin32==311
psycopg2-binary==2.9.11
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63